+++ core/impl/src/main/java/com/blazebit/persistence/impl/AbstractCommonQueryBuilder.java
@@ -248,7 +248,7 @@ public abstract class AbstractCommonQueryBuilder<QueryResultType, BuilderType, S
 
         this.aliasManager = new AliasManager(queryContext.getParent().aliasManager);
         this.expressionFactory = builder.expressionFactory;
-        this.queryGenerator = new ResolvingQueryGenerator(cbf.getMetamodel(), this.aliasManager, parameterManager, mainQuery.parameterTransformerFactory, mainQuery.jpaProvider, mainQuery.registeredFunctions);
+        this.queryGenerator = new ResolvingQueryGenerator(cbf.getMetamodel(), this.aliasManager, parameterManager, mainQuery.parameterTransformerFactory, mainQuery.jpaProvider, mainQuery.dbmsDialect, mainQuery.registeredFunctions);
         this.joinManager = new JoinManager(mainQuery, this, queryGenerator, this.aliasManager, queryContext.getParent().joinManager, expressionFactory);
         this.fromClassExplicitlySet = builder.fromClassExplicitlySet;
 
@@ -301,7 +301,7 @@ public abstract class AbstractCommonQueryBuilder<QueryResultType, BuilderType, S
 
         this.aliasManager = new AliasManager(aliasManager);
         this.expressionFactory = expressionFactory;
-        this.queryGenerator = new ResolvingQueryGenerator(cbf.getMetamodel(), this.aliasManager, parameterManager, mainQuery.parameterTransformerFactory, mainQuery.jpaProvider, mainQuery.registeredFunctions);
+        this.queryGenerator = new ResolvingQueryGenerator(cbf.getMetamodel(), this.aliasManager, parameterManager, mainQuery.parameterTransformerFactory, mainQuery.jpaProvider, mainQuery.dbmsDialect, mainQuery.registeredFunctions);
         this.joinManager = new JoinManager(mainQuery, this, queryGenerator, this.aliasManager, parentJoinManager, expressionFactory);
 
         if (implicitFromClause) {
@@ -1672,7 +1672,7 @@ public abstract class AbstractCommonQueryBuilder<QueryResultType, BuilderType, S
     public BuilderType join(String path, String alias, JoinType type) {
         prepareForModification(ClauseType.JOIN);
         checkJoinPreconditions(path, alias, type);
-        joinManager.join(path, alias, type, false, false);
+        joinManager.join(path, alias, type, false, false, null);
         return (BuilderType) this;
     }
 
@@ -1680,7 +1680,7 @@ public abstract class AbstractCommonQueryBuilder<QueryResultType, BuilderType, S
     public BuilderType joinDefault(String path, String alias, JoinType type) {
         prepareForModification(ClauseType.JOIN);
         checkJoinPreconditions(path, alias, type);
-        joinManager.join(path, alias, type, false, true);
+        joinManager.join(path, alias, type, false, true, null);
         return (BuilderType) this;
     }